init vote extension handlers
