+++ abci/vote_extensions.go
@@ -1,0 +1,19 @@
+package abci
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+// ExtendVoteHandler returns the ExtendVoteHandler ABCI handler that extracts the top bidding valid auction transaction
+// from a validator's local mempool and returns it in its vote extension.
+func (h *ProposalHandler) ExtendVoteHandler() ExtendVoteHandler {
+	return func(ctx sdk.Context, req *RequestExtendVote) (*ResponseExtendVote, error) {
+		panic("implement me")
+	}
+}
+
+// VerifyVoteExtensionHandler returns the VerifyVoteExtensionHandler ABCI handler that verifies the vote extension
+// returned by the ExtendVoteHandler. In particular, it verifies that the vote extension is a valid auction transaction.
+func (h *ProposalHandler) VerifyVoteExtensionHandler() VerifyVoteExtensionHandler {
+	return func(ctx sdk.Context, req *RequestVerifyVoteExtension) (*ResponseVerifyVoteExtension, error) {
+		panic("implement me")
+	}
+}
